+++ tests/frontends/mpd/protocol/music_db_test.py
@@ -13,7 +13,65 @@ class MusicDatabaseHandlerTest(protocol.BaseTestCase):
         self.assertInResponse('OK')
 
     def test_findadd(self):
-        self.sendRequest('findadd "album" "what"')
+        self.backend.library.dummy_find_exact_result = [
+            Track(uri='dummy:a', name='A'),
+        ]
+        self.assertEqual(self.core.tracklist.length.get(), 0)
+
+        self.sendRequest('findadd "title" "A"')
+
+        self.assertEqual(self.core.tracklist.length.get(), 1)
+        self.assertEqual(self.core.tracklist.tracks.get()[0].uri, 'dummy:a')
+        self.assertInResponse('OK')
+
+    def test_searchadd(self):
+        self.backend.library.dummy_search_result = [
+            Track(uri='dummy:a', name='A'),
+        ]
+        self.assertEqual(self.core.tracklist.length.get(), 0)
+
+        self.sendRequest('searchadd "title" "a"')
+
+        self.assertEqual(self.core.tracklist.length.get(), 1)
+        self.assertEqual(self.core.tracklist.tracks.get()[0].uri, 'dummy:a')
+        self.assertInResponse('OK')
+
+    def test_searchaddpl_appends_to_existing_playlist(self):
+        playlist = self.core.playlists.create('my favs').get()
+        playlist = playlist.copy(tracks=[
+            Track(uri='dummy:x', name='X'),
+            Track(uri='dummy:y', name='y'),
+        ])
+        self.core.playlists.save(playlist)
+        self.backend.library.dummy_search_result = [
+            Track(uri='dummy:a', name='A'),
+        ]
+        playlists = self.core.playlists.filter(name='my favs').get()
+        self.assertEqual(len(playlists), 1)
+        self.assertEqual(len(playlists[0].tracks), 2)
+
+        self.sendRequest('searchaddpl "my favs" "title" "a"')
+
+        playlists = self.core.playlists.filter(name='my favs').get()
+        self.assertEqual(len(playlists), 1)
+        self.assertEqual(len(playlists[0].tracks), 3)
+        self.assertEqual(playlists[0].tracks[0].uri, 'dummy:x')
+        self.assertEqual(playlists[0].tracks[1].uri, 'dummy:y')
+        self.assertEqual(playlists[0].tracks[2].uri, 'dummy:a')
+        self.assertInResponse('OK')
+
+    def test_searchaddpl_creates_missing_playlist(self):
+        self.backend.library.dummy_search_result = [
+            Track(uri='dummy:a', name='A'),
+        ]
+        self.assertEqual(
+            len(self.core.playlists.filter(name='my favs').get()), 0)
+
+        self.sendRequest('searchaddpl "my favs" "title" "a"')
+
+        playlists = self.core.playlists.filter(name='my favs').get()
+        self.assertEqual(len(playlists), 1)
+        self.assertEqual(playlists[0].tracks[0].uri, 'dummy:a')
         self.assertInResponse('OK')
 
     def test_listall(self):